Update everything and improve hardware support
Huge changes!

+++ board/common/post-build.sh
@@ -1,20 +1,23 @@
 #!/bin/sh
-# Información de versión más apropiada que la predeterminada
-printf "NAME=\"Linux\"\nVERSION=\"5.3.1\"\nID=linux\nID_LIKE=\"buildroot\"\nHOME_URL=\"https://github.com/pinetainer\"\nVARIANT=\"PineTainer\"\nVARIANT_ID=\"PineTainer\"" > "$1/usr/lib/os-release"
 
-# Borrar ficheros inútiles solo usados por bash
-rm -rf "$1/etc/bash_completion.d" || true
+# Tweak urandom service start order
+mv "${1:?}/etc/init.d/S20urandom" "${1:?}/etc/init.d/S01urandom" 2>/dev/null || true
 
-# Hacer que urandom sea de los primeros servicios en ser iniciado,
-# para tener acceso a mayor aleatoriedad antes
-mv "$1/etc/init.d/S20urandom" "$1/etc/init.d/S01urandom.sh" || true
+# Remove rmt, which is a executable used by tar only for
+# legacy devices
+rm "${1:?}/libexec/rmt" 2>/dev/null || true
+rmdir "${1:?}/libexec" 2>/dev/null || true
 
-# Borrar ficheros que cuelgan de directorios temporales
-find "$1/tmp" -mindepth 1 -delete || true
-find "$1/run" -mindepth 1 -delete || true
-find "$1/dev" -mindepth 1 -delete || true
-find "$1/var/cache" -mindepth 1 -delete || true
+# Remove LVM configuration files which take 100 KiB of space
+rm "${1:?}/etc/lvm/lvm.conf" 2>/dev/null || true
+rm "${1:?}/etc/lvm/lvmlocal.conf" 2>/dev/null || true
 
-# Borrar rmt y el directorio que lo contiene, si queda vacío
-rm -f "$1/libexec/rmt" || true
-rmdir "$1/libexec" 2>/dev/null || true
+# Save ~200 MiB in useless files
+for dri_lib in \
+armada-drm_dri.so hx8357d_dri.so ili9341_dri.so ingenic-drm_dri.so meson_dri.so \
+mxsfb-drm_dri.so pl111_dri.so rockchip_dri.so st7735r_dri.so exynos_dri.so \
+ili9225_dri.so imx-drm_dri.so mcde_dri.so mi0283qt_dri.so repaper_dri.so \
+st7586_dri.so stm_dri.so
+do
+    rm "${1:?}/usr/lib/dri/$dri_lib" || true
+done
